python/django/ptc_deco/requirements.txt
aniso8601==7.0.0
appdirs==1.4.4
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
backcall==0.2.0
bcrypt==3.2.0
black==20.8b1
brunette==0.2.0
cffi==1.14.5
click==7.1.2
colorama==0.4.4
cryptography==3.4.7
dataclasses-json==0.5.2
DAWG==0.8.0
DAWG-Python==0.7.2
decorator==4.4.2
Django==3.1.7
django-cors-headers==3.7.0
django-currentuser==0.5.3
django-extensions==3.1.1
django-filter==2.4.0
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
django-print-sql==2018.3.6
django-simple-history==3.0.0
djangorestframework==3.12.3
docopt==0.6.2
et-xmlfile==1.0.1
fabric==2.5.0
graphene-django==2.15.0
graphene-django-auth==0.1.1
graphene-django-cud==0.7.2
graphene-file-upload==1.3.0
graphql-core==2.3.2
gunicorn==20.0.4
iniconfig==1.1.1
invoke==1.4.1
ipython==7.21.0
ipython-genutils==0.2.0
iso8601==0.1.14
jedi==0.18.0
loguru==0.5.3
lxml==4.6.3
marshmallow==3.11.1
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
openpyxl==3.0.7
packaging==20.9
paramiko==2.7.2
parso==0.8.1
patchwork==1.0.1
pathspec==0.8.1
pickleshare==0.7.5
Pillow==8.1.2
pluggy==0.13.1
promise==2.3
prompt-toolkit==3.0.18
protobuf==3.15.6
psycopg2==2.8.6
psycopg2-binary==2.8.6
py==1.10.0
pycparser==2.20
pydevd-pycharm==201.7223.92
Pygments==2.8.1
PyJWT==1.7.1
pymorphy2==0.9.1
pymorphy2-dicts==2.4.393442.3710985
pymorphy2-dicts-ru==2.4.417127.4579844
PyNaCl==1.4.0
pyparsing==2.4.7
pytest==6.1.2
python-dateutil==2.8.1
python-docx==0.8.10
python-dotenv==0.15.0
pytz==2021.1
PyYAML==5.3.1
recordtype==1.3
regex==2021.3.17
Rx==1.6.1
singledispatch==3.6.1
six==1.15.0
sqlparse==0.4.1
stringcase==1.2.0
text-unidecode==1.3
toml==0.10.2
tqdm==4.59.0
traitlets==5.0.5
typed-ast==1.4.2
typing-extensions==3.7.4.3
typing-inspect==0.6.0
wcwidth==0.2.5
django-graphql-jwt<=0.3.0
graphene @ git+git://github.com/ptcfrontend/graphene@total-count#egg=graphene
graphql_relay @ git+git://github.com/ptcfrontend/graphql-relay-py@total-count#egg=graphql_relay
pydevd-pycharm~=201.7223.92
django-postgres-extra
django-ipware~=3.0.2
django-axes~=5.19.0
# pypiwin32
# python venv\Scripts\pywin32_postinstall.py -install